change interface for other EC2 service
